checker add

+++ list_utils.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   list_utils.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: medayi    <medayi@student.42istanbul>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/03/09 06:11:03 by medayi            #+#    #+#             */
+/*   Updated: 2025/03/09 06:27:30 by medayi           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "push_swap.h"
 
 t_list  *ft_lstnew(int data)
@@ -12,7 +24,7 @@ t_list  *ft_lstnew(int data)
     return (new);
 }
 
-int ft_lstsize(t_list   *lst)
+size_t ft_lstsize(t_list   *lst)
 {
     size_t  size;
